requirements.txt
Django==3.0.3
djangorestframework==3.11.0
django-filter==2.3.0
django-rest-auth==0.9.5
psycopg2==2.8.5
Markdown==3.2.2